Implement Phase 1 Enhanced UX - Backend Features
This commit implements the core backend features for Phase 1 of the roadmap:

### 🔧 Bulk Operations API
- Add GraphQL mutations: bulkUpdateTasks, bulkDeleteTasks, bulkAssignUsers
- Support operations on multiple task IDs with validation
- Add BulkOperationResult type for consistent error handling
- Include audit logging through WebSocket notifications

### 🔄 WebSocket Integration
- Add Spring WebSocket support with STOMP messaging
- Implement WebSocketNotificationService for real-time updates
- Real-time notifications for task create/update/delete operations
- Support for bulk operation broadcasts
- User presence and assignment notifications

### 📋 Task Templates System
- New TaskTemplate entity with name, title, description, estimationHours
- Complete CRUD operations via GraphQL API
- Template-based task creation with user assignment
- Template name uniqueness validation
- Integration with existing task workflow

### 🏗️ Infrastructure Updates
- Add spring-boot-starter-websocket dependency
- Extend GraphQL schema with new types and mutations
- Fix type consistency issues (Integer -> Double for estimation hours)
- Add proper validation and error handling
- Maintain backwards compatibility with existing API

All backend features are ready for frontend integration in subsequent commits.

+++ src/main/java/com/projects/taskmanager/repository/TaskRepository.java
@@ -47,8 +47,8 @@ public interface TaskRepository extends JpaRepository<Task, Long> {
         @Param("status") TaskStatus status,
         @Param("dueDateFrom") LocalDate dueDateFrom,
         @Param("dueDateTo") LocalDate dueDateTo,
-        @Param("estimationHoursMin") Integer estimationHoursMin,
-        @Param("estimationHoursMax") Integer estimationHoursMax,
+        @Param("estimationHoursMin") Double estimationHoursMin,
+        @Param("estimationHoursMax") Double estimationHoursMax,
         Pageable pageable
     );
